Remove console logs, change alert messages, and move api call files
