--- layout/main.hbs
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>Santa's Secret</title>
-  <link rel="stylesheet" href="normalise.css">
-  <link rel="stylesheet" href="style.css">
-  <link href="https://fonts.googleapis.com/css?family=Jaldi:700" rel="stylesheet">
-</head>
-<body>
-
-  {{> header}}
-
-  {{{body}}}
-
-  {{> footer}}
-</body>
-
-</html>
